Adding ModalTextInput and moving tailwind files to scoped css
The tailwind.out.css file in global scope when added with link
element was causing problem in the main app integrating this
library component, due to duplicate classes in the tailwind.out.css
file in the main app itself.
Hence moved the tailwind files to individual component folder,
where it will be scoped to the component.