Moving tailwind compilation of components to the parent
The tailwind classes if compiled with each component result in
a lot of redundancy and increased css download size.
So to optimize the overall css download size, removing all taiwind
compilation from individual utility components, and adding the
compilation in the parent project including these components.

The including project must make sure to add these component paths
in its own tailwind.config.js so that the classes in use here
are compiled and loaded in the parent project where these utility
components are going to be used.